% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl.R
\name{adjustedRandIndex}
\alias{adjustedRandIndex}
\title{Adjusted Rand Index}
\usage{
adjustedRandIndex(x, y)
}
\arguments{
\item{x}{a vector of classes}

\item{y}{a vector of classes}
}
\description{
Compute the Adjusted Rand Index for two set of classes
}
\examples{
A = c(1,1,1,2,2,2)
B = c(1,2,1,2,2,1)
adjustedRandIndex(A,B)
}
