% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl.R
\name{generateDataset}
\alias{generateDataset}
\title{Generate a time series dataset}
\usage{
generateDataset()
}
\value{
\describe{
\item{X}{a matrix of 300 time series (300 rows and 256 columns)}
\item{col}{a vector, describing the class of each row (i.e. curve)}
}
}
\description{
The dataset composed by 300 curves of 256 time points, from 3 different groups.
}
