% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl.R
\name{plot.curve.dataset}
\alias{plot.curve.dataset}
\title{Plot a dataset of curves in a user friendly way}
\usage{
\method{plot}{curve.dataset}(X, col = rep(1, nrow(X)), ratio = 1)
}
\arguments{
\item{X}{a matrix of time series (one time serie per line of the matrix)}

\item{col}{the class vector of the data (must be numbers)}

\item{ratio}{the ratio of curves plotted (1 : all curves plotted, 0.25 : 25\% of the curves plotted)}
}
\description{
Plot the curves of the dataset \emph{X}, according to the curve's class, provided by the \emph{col} parameter
With the \emph{ratio} parameter, you can reduce the number of plotted curves (ex : ratio = 0.1 plot only 10% of the curves).
}
\examples{
\dontrun{
dataset = generateDataset()
plot.curve.dataset(X=dataset$X,col=dataset$col,ratio=0.1)
}

}
