% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl.R
\name{plot.mean.curve.dataset}
\alias{plot.mean.curve.dataset}
\title{Plot the mean curves of the dataset}
\usage{
\method{plot}{mean.curve.dataset}(X, col = rep(1, nrow(X)))
}
\arguments{
\item{X}{a matrix of time series (one time serie per line of the matrix)}

\item{col}{the class vector of the data (must be numbers)}
}
\description{
For each group of curves, plot the mean of the curve, and 10 curves from the plot.
}
